requirements.txt
Django 
djangorestframework
psycopg2-binary
python-decouple
Pillow
drf-yasg
django-filter